requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4
google-generativeai==0.8.3
python-multipart==0.0.20
dateparser==1.2.0